refactor(/home): reorganize files
