+++ ourportfolios/pages/home/components/__init__.py
@@ -1,7 +1,11 @@
 """Composable Home page components."""
 
-from .framework_card import select_framework_card, selected_framework_card
-from .hub_cards import compare_assets_card, manage_portfolio_card
+from .hub_cards import (
+    compare_assets_card,
+    manage_portfolio_card,
+    select_framework_card,
+    selected_framework_card,
+)
 
 __all__ = [
     "select_framework_card",
