+++ ourportfolios/pages/home/components/refresh_countdown.py
@@ -1,8 +1,8 @@
 """Refresh countdown rendered directly from HomeState."""
 
 import reflex as rx
-from ...state.home_state import HomeState
-from ...styles import white
+from ....state.home_state import HomeState
+from ....styles import white
 
 _CIRC = 54.85
