+++ ourportfolios/pages/home/components/ticker_of_day.py
@@ -1,8 +1,8 @@
 import reflex as rx
-from ...state.home_state import HomeState
-from ...state.cart_state import CartState
-from ...components.cards import glass_card
-from ...styles import white
+from ....state.home_state import HomeState
+from ....state.cart_state import CartState
+from ....components.cards import glass_card
+from ....styles import white
 
 
 def _skel(w: str = "100%", h: str = "0.75rem", r: str = "0.375rem") -> rx.Component:
