--- ourportfolios/pages/home/framework_card.py
@@ -1,3 +1,0 @@
-from .components.framework_card import select_framework_card, selected_framework_card
-
-__all__ = ["select_framework_card", "selected_framework_card"]
