--- ourportfolios/pages/home/hub_cards.py
@@ -1,3 +1,0 @@
-from .components.hub_cards import compare_assets_card, manage_portfolio_card
-
-__all__ = ["compare_assets_card", "manage_portfolio_card"]
