--- ourportfolios/pages/home/page_body.py
@@ -1,3 +1,0 @@
-from .sections.dashboard import page_body
-
-__all__ = ["page_body"]
