+++ ourportfolios/state/heatmap.py
@@ -21,7 +21,7 @@ MAX_TREEMAP_INDUSTRIES = 9
 ITEMS_PER_ROW = 3
 _CTR_W = 760.0
 _CTR_H = 620.0
-LABEL_H_PX = 30.0
+LABEL_H_PX = 36.0
 _MIN_PX = 52.0
 _MAX_N = 4
 _MIN_TILE_H_PX = 80.0
